requirements.txt
aiofiles==0.4.0
aiologger==0.6.0
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
asn1crypto==1.4.0
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.3
bcrypt==3.2.0
boto3==1.14.27
botocore==1.17.27
btrfsutil==5.11.1
CacheControl==0.12.6
cached-property==1.5.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==3.4.7
cupshelpers==1.0
dict-toolbox==1.12
distlib==0.3.1
distro==1.5.0
dnspython==1.16.0
docker==4.4.4
docker-compose==1.28.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
evdev==1.4.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
html5lib==1.1
idna==3.1
importlib-metadata==3.7.3
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
keyutils==0.6
lazy-object-proxy==1.4.3
lensfun==0.3.95
LibAppArmor==3.0.1
lml==0.0.9
louis==3.17.0
lutris==0.5.8.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
meson==0.57.1
msgpack==1.0.0
netsnmp-python==1.0a1
new-flask-app==1.0.2
npyscreen==4.10.5
numpy==1.19.1
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.0
pandas==1.2.4
paramiko==2.7.2
pep517==0.10.0
Pillow==8.1.2
pluggy==0.13.1
ply==3.11
pop==17.1
pop-config==6.11
progress==1.5
prompt-toolkit==1.0.14
py==1.10.0
pybind11==2.6.2
pycairo==1.20.0
pycodestyle==2.6.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
pyexcel==0.6.4
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
Pygments==2.8.1
PyGObject==3.40.0
PyInquirer==1.0.3
pylint==2.5.3
PyMySQL==0.10.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyrsistent==0.17.3
pysmbc==1.0.23
pytest==6.2.3
pytest-mock==3.5.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-libtorrent==1.2.11
pytz==2020.1
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
regex==2021.4.4
reportlab==3.5.66
requests==2.24.0
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.6.2
six==1.15.0
SQLAlchemy==1.3.18
team==1.0
texttable==1.6.2
threadpoolctl==2.1.0
toml==0.10.1
tox==3.23.0
typed-ast==1.4.1
udiskie==2.3.3
urllib3==1.25.10
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.3
zipp==3.4.1
